feat: add internal plugin metrics
The plugin can now emit metrics describing its internal state. These
describe the flow of data through the plugin in terms of the number of
records and batches that were either sent or dropped, as well as buffer
sizes. The metrics are disabled by default, and can be enabled by adding
`EnableInternalMetrics = true` to the configuration.